+++ src/main/java/org/sonar/plugins/delphi/type/DelphiType.java
@@ -5,19 +5,23 @@ import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.ANSISTRING;
 import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.CHAR;
 import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.SHORTSTRING;
 import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.UNICODESTRING;
-import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.WIDECHAR;
 import static org.sonar.plugins.delphi.type.intrinsic.IntrinsicText.WIDESTRING;
 
+import com.google.errorprone.annotations.Immutable;
 import java.util.Collections;
 import java.util.Set;
 import org.jetbrains.annotations.NotNull;
 import org.sonar.plugins.delphi.type.generic.TypeSpecializationContext;
 
 public abstract class DelphiType implements Type {
-  private final String image;
+  private final ImageSupplier<?> imageSupplier;
 
   protected DelphiType(String image) {
-    this.image = image;
+    this.imageSupplier = type -> image;
+  }
+
+  protected <T extends DelphiType> DelphiType(ImageSupplier<?> imageSupplier) {
+    this.imageSupplier = imageSupplier;
   }
 
   public static ImmutableType unknownType() {
@@ -34,7 +38,7 @@ public abstract class DelphiType implements Type {
 
   @Override
   public String getImage() {
-    return image;
+    return imageSupplier.getImage(this);
   }
 
   @NotNull
@@ -256,5 +260,16 @@ public abstract class DelphiType implements Type {
   @Override
   public String toString() {
     return getImage();
+  }
+
+  @FunctionalInterface
+  @Immutable
+  protected interface ImageSupplier<T extends Type> {
+    @SuppressWarnings("unchecked")
+    default String getImage(DelphiType type) {
+      return getImage((T) type);
+    }
+
+    String getImage(T type);
   }
 }